pyproject.toml
[project]
name = "sugarbeet-weed-detection"
version = "0.1.0"
description = "Weed detection in sugarbeet fields"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lightning==2.6.1",
    "numpy>=2.4.2",
    "pillow>=12.1.0",
    "pycocotools>=2.0.11",
    "sahi>=0.11.36",
    "torch==2.6.0",
    "ultralytics>=8.4.14",
    "wandb>=0.24.2",
]
